Refactor helper to accept slots by clinic id
