+++ src/applications/vaos/tests/new-appointment/components/DateTimeSelectPage/helpers.js
@@ -4,15 +4,14 @@ import {
   mockAppointmentSlotFetch,
   mockEligibilityFetches,
 } from '../../../mocks/helpers';
-import { getAppointmentSlotMock, getClinicMock } from '../../../mocks/v0';
+import { getClinicMock } from '../../../mocks/v0';
 
 export function setDateTimeSelectMockFetches({
   preferredDate = moment(),
-  singleClinic = false,
-  singleSlot = false,
   slotError = false,
-  slotsToSet = [],
+  slotDatesByClinicId = {},
 } = {}) {
+  const clinicIds = Object.keys(slotDatesByClinicId);
   const clinics = [
     {
       id: '308',
@@ -43,56 +42,18 @@ export function setDateTimeSelectMockFetches({
     limit: true,
     requestPastVisits: true,
     directPastVisits: true,
-    clinics: singleClinic ? [clinics[0]] : clinics,
+    clinics: clinicIds.length === 2 ? clinics : [clinics[0]],
     pastClinics: true,
   });
   if (!slotError) {
-    const slot308Date = moment()
-      .day(9)
-      .hour(9)
-      .minute(0)
-      .second(0);
-    const slot309Date = moment()
-      .day(11)
-      .hour(13)
-      .minute(0)
-      .second(0);
-
-    mockAppointmentSlotFetch({
-      siteId: '983',
-      clinicId: '308',
-      typeOfCareId: '323',
-      slots: slotsToSet.length
-        ? slotsToSet
-        : [
-            {
-              ...getAppointmentSlotMock(),
-              startDateTime: slot308Date.format('YYYY-MM-DDTHH:mm:ss[+00:00]'),
-              endDateTime: slot308Date
-                .clone()
-                .minute(20)
-                .format('YYYY-MM-DDTHH:mm:ss[+00:00]'),
-            },
-          ],
-      preferredDate,
-    });
-    if (!singleSlot) {
+    clinicIds.forEach(id => {
       mockAppointmentSlotFetch({
         siteId: '983',
-        clinicId: '309',
+        clinicId: id,
         typeOfCareId: '323',
-        slots: [
-          {
-            ...getAppointmentSlotMock(),
-            startDateTime: slot309Date.format('YYYY-MM-DDTHH:mm:ss[+00:00]'),
-            endDateTime: slot309Date
-              .clone()
-              .minute(20)
-              .format('YYYY-MM-DDTHH:mm:ss[+00:00]'),
-          },
-        ],
+        slots: slotDatesByClinicId[id],
         preferredDate,
       });
-    }
+    });
   }
 }
